% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-bc-species.R
\name{combine_bc_species}
\alias{combine_bc_species}
\title{Combine BC Species and DB Species}
\usage{
combine_bc_species(bc_species, db_species)
}
\arguments{
\item{bc_species}{A data frame}

\item{db_species}{A data frame}
}
\value{
A data frame
}
\description{
Internal to allow for testing
}
\examples{
\dontrun{
data <- combine_bc_species(bc_species, db_species)
}
}
