% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-bc-wqg.R
\name{combine_bc_wqg}
\alias{combine_bc_wqg}
\title{Combine BC Water Quality Guidelines with DB Chemicals}
\usage{
combine_bc_wqg(bc_wqg, db_chemicals)
}
\arguments{
\item{bc_wqg}{A data frame}

\item{db_chemicals}{A data frame}
}
\value{
A data frame
}
\description{
Internal to allow for testing
}
\examples{
\dontrun{
data <- combine_bc_wqg(bc_wqg, db_chemicals)
}
}
