% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-conc-conversions.R
\name{combine_conc_conversions}
\alias{combine_conc_conversions}
\title{Combine Concentration Unit Conversation Values with DB Concentration Unit}
\usage{
combine_conc_conversions(concentration_std, db_concentration_unit_codes)
}
\arguments{
\item{concentration_std}{A data frame}

\item{db_concentration_unit_codes}{A data frame}
}
\value{
A data frame
}
\description{
Internal to allow for testing
}
\examples{
\dontrun{
data <- combine_conc_conversions(
  concentration_std, db_concentration_unit_codes
)
}
}
