% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-concentration-endpoints.R
\name{combine_concentration_endpoints}
\alias{combine_concentration_endpoints}
\title{Combine the Concentration Endpoints list to db endpoints}
\usage{
combine_concentration_endpoints(endpoint_concentration_pick, db_endpoint_code)
}
\arguments{
\item{endpoint_concentration_pick}{A data frame}

\item{db_endpoint_code}{A data frame}
}
\value{
A data frame
}
\description{
Internal to allow for testing
}
\examples{
\dontrun{
data <- combine_concentration_endpoints(
 endpoint_concentration_pick, db_endpoint_code
)
}
}
