% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-duration-conversions.R
\name{combine_duration_conversions}
\alias{combine_duration_conversions}
\title{Combine Duration Unit Conversation Values and DB duration unit}
\usage{
combine_duration_conversions(duration_std, db_duration_unit_codes)
}
\arguments{
\item{duration_std}{A data frame}

\item{db_duration_unit_codes}{A data frame}
}
\value{
A data frame
}
\description{
Internal to allow for testing
}
\examples{
\dontrun{
duration_unit_codes_std <- combine_duration_conversions(
  duration_std, db_duration_unit_codes
)
}
}
