% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-lifestage.R
\name{combine_lifestage}
\alias{combine_lifestage}
\title{Combine Life Stage Groups and db life stage}
\usage{
combine_lifestage(lifestage_codes, db_lifestage_codes)
}
\arguments{
\item{lifestage_codes}{A data frame}

\item{db_lifestage_codes}{A data frame}
}
\value{
A data frame
}
\description{
Internal to allow for testing
}
\examples{
\dontrun{
lifestage_groups <- combine_lifestage(lifestage_codes, db_lifestage_codes)
}
}
