% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-media.R
\name{combine_media}
\alias{combine_media}
\title{Combine Media Type Coding}
\usage{
combine_media(db_media_type)
}
\arguments{
\item{db_media_type}{A data frame}
}
\value{
A data frame
}
\description{
Internal to allow for testing
}
\examples{
\dontrun{
media_type <- combine_media(db_media_type)
}
}
