% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-trophic-group.R
\name{combine_trophic_group}
\alias{combine_trophic_group}
\title{Combine Trophic Groups and DB species}
\usage{
combine_trophic_group(trophic_groups, db_species)
}
\arguments{
\item{trophic_groups}{A data frame}

\item{db_species}{A data frame}
}
\value{
A data frame
}
\description{
Internal to allow for testing
}
\examples{
\dontrun{
species_trophic_group <- combine_trophic_group(trophic_groups, db_species)
}
}
