% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-join-data.R
\name{join_data}
\alias{join_data}
\title{Join Data}
\usage{
join_data(
  db_results,
  db_tests,
  db_endpoint_codes,
  db_species,
  db_lifestage_codes,
  db_chemicals,
  db_duration_unit_codes,
  db_concentration_unit_codes,
  db_references,
  db_effect_codes,
  db_media_type_codes,
  db_meta_data_download
)
}
\arguments{
\item{db_results}{A data frame}

\item{db_tests}{A data frame}

\item{db_endpoint_codes}{A data frame}

\item{db_species}{A data frame}

\item{db_lifestage_codes}{A data frame}

\item{db_chemicals}{A data frame}

\item{db_duration_unit_codes}{A data frame}

\item{db_concentration_unit_codes}{A data frame}

\item{db_references}{A data frame}

\item{db_effect_codes}{A data frame}

\item{db_media_type_codes}{A data frame}

\item{db_meta_data_download}{A data frame}
}
\value{
Invisible data frame
}
\description{
Internal to wqb_join_data() to allow for testing.
}
\examples{
\dontrun{
data <- wqb_join_data(
 db_results, db_tests, db_endpoint_codes, db_species, db_lifestage_codes, 
 db_chemicals, db_duration_unit_codes, db_concentration_unit_codes, 
 db_references, db_effect_codes, db_media_type_codes, db_meta_data_download
) 
}
}
