% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-bc-species.R
\name{read_bc_species}
\alias{read_bc_species}
\title{Read in BC Species CSV}
\usage{
read_bc_species(bc_species_file_path, db_species)
}
\arguments{
\item{bc_species_file_path}{A file path}

\item{db_species}{A data frame}
}
\value{
A data frame
}
\description{
Internal to allow for testing
}
\examples{
\dontrun{
data <- read_bc_species(bc_species_file_path, db_species)
}
}
