% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-bc-wqg.R
\name{read_bc_wqg}
\alias{read_bc_wqg}
\title{Read BC Water Quality Guidelines}
\usage{
read_bc_wqg(db_chemicals)
}
\arguments{
\item{db_chemicals}{A data frame}
}
\value{
A data frame
}
\description{
Internal to allow for testing
}
\examples{
\dontrun{
chemicals_bc_wqg <- read_bc_wqg(db_chemicals)
}
}
