% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-conc-conversions.R
\name{read_conc_conversions}
\alias{read_conc_conversions}
\title{Read Concentration Unit Conversation Values}
\usage{
read_conc_conversions(concentration_std_file_path, db_concentration_unit_codes)
}
\arguments{
\item{concentration_std_file_path}{A file path}

\item{db_concentration_unit_codes}{A data frame}
}
\value{
A data frame
}
\description{
Internal to allow for testing
}
\examples{
\dontrun{
concentration_unit_codes_std <- read_conc_conversions(
  concentration_std_file_path, db_concentration_unit_codes
)
}
}
