% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-concentration-endpoints.R
\name{read_concentration_endpoints}
\alias{read_concentration_endpoints}
\title{Read the Concentration Endpoints}
\usage{
read_concentration_endpoints(conc_endpoints_file_path, db_endpoint_code)
}
\arguments{
\item{conc_endpoints_file_path}{A file path}

\item{db_endpoint_code}{A data frame}
}
\value{
A data frame
}
\description{
Internal to allow for testing
}
\examples{
\dontrun{
endpoint_concentration <- read_concentration_endpoints(
  conc_endpoints_file_path, db_endpoint_code
)
}
}
