% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-duration-conversions.R
\name{read_duration_conversions}
\alias{read_duration_conversions}
\title{Read Duration Unit Conversation}
\usage{
read_duration_conversions(duration_std_file_path, db_duration_unit_codes)
}
\arguments{
\item{duration_std_file_path}{A file path}

\item{db_duration_unit_codes}{A data frame}
}
\value{
A data frame
}
\description{
Internal to allow for testing
}
\examples{
\dontrun{
duration_unit_codes_std <- read_duration_conversions(
  duration_std_file_path, db_duration_unit_codes
)
}
}
