% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-lifestage.R
\name{read_lifestage}
\alias{read_lifestage}
\title{Read Life Stage Groups}
\usage{
read_lifestage(lifestage_file_path, db_lifestage_codes)
}
\arguments{
\item{lifestage_file_path}{A data frame}

\item{db_lifestage_codes}{A data frame}
}
\value{
A data frame
}
\description{
Internal to allow for testing
}
\examples{
\dontrun{
lifestage_groups <- read_lifestage(lifestage_file_path, db_lifestage_codes)
}
}
