% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-trophic-group.R
\name{read_trophic_group}
\alias{read_trophic_group}
\title{Read Trophic Groups and DB species}
\usage{
read_trophic_group(trophic_groups_file_path, db_species)
}
\arguments{
\item{trophic_groups_file_path}{A file path}

\item{db_species}{A data frame}
}
\value{
A data frame
}
\description{
Internal to allow for testing
}
\examples{
\dontrun{
species_trophic_group <- read_trophic_group(trophic_groups_file_path, db_species)
}
}
