% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-bc-wqg.R
\name{wqb_add_bc_wqg}
\alias{wqb_add_bc_wqg}
\title{Add BC Water Quality Guidelines}
\usage{
wqb_add_bc_wqg(database, quiet = FALSE)
}
\arguments{
\item{database}{A string to the location of the database.}

\item{quiet}{Turn off message when quiet set to TRUE.}
}
\value{
Invisible data frame
}
\description{
Read in the British Columbia water quality guidelines (wqg) and add a column
to the chemicals tables in the database to indicate if the chemical is
present in the British Columbia water quality guidelines.
}
\details{
The wqg data is stored in the BC Data Catalogue.

The \code{CAS_number} column in the bc wqg data is matched to the \code{cas_number}
column in the chemicals table of the database. A new column
\code{present_in_bc_wqg} is added to the chemicals table that codes each
chemical as TRUE if the chemical is present in wqg or FALSE if the chemical
is not present in wqg.
}
\examples{
\dontrun{
chem_bc_wqg <- wqb_add_bc_wqg(
  database = "ecotox_ascii_09_15_2022.sqlite"
)

chem_bc_wqg <- wqb_add_bc_wqg(
  database = "ecotox_db/ecotox_ascii_09_15_2022.sqlite"
)
}
}
