% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-lifestage.R
\name{wqb_add_lifestage}
\alias{wqb_add_lifestage}
\title{Add corrected Life stage codes to Database}
\usage{
wqb_add_lifestage(database)
}
\arguments{
\item{database}{A string to the location of the database.}
}
\value{
Invisible data frame
}
\description{
The corresponding life stage codes for fish and amphibians.
}
\examples{
\dontrun{
lifestage_codes <- wqb_add_lifestage(
 database = "ecotox_ascii_09_15_2022.sqlite"
) 

lifestage_codes <- wqb_add_lifestage(
 database = "ecotox_db/ecotox_ascii_09_15_2022.sqlite"
) 
}
}
