% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-media.R
\name{wqb_add_media}
\alias{wqb_add_media}
\title{Add Media Type Coding of Salt or Fresh Water}
\usage{
wqb_add_media(database, quiet = FALSE)
}
\arguments{
\item{database}{A string to the location of the database.}

\item{quiet}{Turn off message when quiet set to TRUE.}
}
\value{
Invisible data frame
}
\description{
Code the media type information into three categories: salt water, fresh
water and not reported.
}
\details{
Read in the media_type_codes table from the database and code the 24
groups into three categories: salt water, fresh water or not reported. The
new categories are added to a column in the table called
\code{media_type_group}.
}
\examples{
\dontrun{
media_info <- wqb_add_media(
  database = "ecotox_ascii_09_15_2022.sqlite"
)

media_info <- wqb_add_media(
  database = "ecotox_db/ecotox_ascii_09_15_2022.sqlite"
)
}
}
