% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-trophic-group.R
\name{wqb_add_trophic_group}
\alias{wqb_add_trophic_group}
\title{Add Trophic Group to Species in the Database}
\usage{
wqb_add_trophic_group(file_path, database)
}
\arguments{
\item{file_path}{A string to the file that contains the trophic information.}

\item{database}{A string to the location of the database.}
}
\value{
Invisible data frame
}
\description{
Read in and add the trophic groups to the species. Species are classed into
groups and class level groups.
}
\details{
The trophic groups file must contain the columns: class, order,
ecological_group, ecological_group_class. The column names are case
sensitive, ensure they are lower case.

The output table contains three columns: species_number, ecological_group,
and ecological_group_class.

The output table is added to the database with the name
\code{species_trophic_group}.
}
\examples{
\dontrun{
# all files in root directory
trophic_group <- wqb_add_trophic_group(
 file_path = "trophic-group.csv",
 database = "ecotox_ascii_09_15_2022.sqlite"
) 

# files in subdirectories 
trophic_group <- wqb_add_trophic_group(
 file_path = "lookups/trophic-group.csv",
 database = "ecotox_db/ecotox_ascii_09_15_2022.sqlite"
) 
}
}
