% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-af.R
\name{wqb_af}
\alias{wqb_af}
\title{Determine the Assessment Factors}
\usage{
wqb_af(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame
}
\description{
Determines the species variation factor, ecological assessment factor and
B.C. species factor.
}
\details{
This is a wrapper function that calls \code{wqb_af_bc_species()},
\code{wqb_af_ecological()} and \code{wqb_af_variation()}.
}
\examples{
\dontrun{
data <- wqb_af(data)
}
}
