% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-af-bc-species.R
\name{wqb_af_bc_species}
\alias{wqb_af_bc_species}
\title{Determine the BC Species Assessment Factors}
\usage{
wqb_af_bc_species(data)
}
\arguments{
\item{data}{A data frame that is the output of the \code{wqb_benchmark-method()} function.}
}
\value{
A data frame
}
\description{
Determine the BC Species Assessment Factors
}
\examples{
\dontrun{
data <- wqb_af_bc_species(data)
}
}
