% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-af-ecological.R
\name{wqb_af_ecological}
\alias{wqb_af_ecological}
\title{Determine the Ecological Assessment Factors}
\usage{
wqb_af_ecological(data)
}
\arguments{
\item{data}{A data frame that is the output of the \code{wqb_benchmark-method()}
function.}
}
\value{
A data frame
}
\description{
Depending on the presence of certain ecological groups present in the data
set determines the assessment factor applied.
}
\examples{
\dontrun{
data <- wqb_af_ecological(data)
}
}
