% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-af-variation.R
\name{wqb_af_variation}
\alias{wqb_af_variation}
\title{Determine the Variation Assessment Factor}
\usage{
wqb_af_variation(data)
}
\arguments{
\item{data}{A data frame that is the output of the \code{wqb_benchmark-method()}
function.}
}
\value{
A data frame
}
\description{
Depending on the number of trophic groups and species a certain assessment
factor is determined.
}
\examples{
\dontrun{
data <- wqb_af_variation(data)
}
}
