% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-aggregate.R
\name{wqb_aggregate}
\alias{wqb_aggregate}
\title{Aggregate Data for each Species}
\usage{
wqb_aggregate(data, cas_num)
}
\arguments{
\item{data}{A data frame that is the output of the \code{wqb_standardize_effect()} function.}

\item{cas_num}{A string of the cas number.}
}
\value{
Data frame
}
\description{
Aggregate data so the most sensitive value is selected for each species.
}
\details{
Each species, life stage and effect are group for each chemical and the
most sensitive effect concentration is selected for each speices.

If multiple data points within each group exists then the geometric mean is
calculated.
}
\examples{
\dontrun{
wqb_aggregate(data, "71432")
wqb_aggregate(data, "67663")
}
}
