% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-benchmark-method.R
\name{wqb_benchmark_method}
\alias{wqb_benchmark_method}
\title{Determine Benchmark Method}
\usage{
wqb_benchmark_method(data)
}
\arguments{
\item{data}{A data frame that is the output of the \code{wqb_aggregate()} function.}
}
\value{
A data frame
}
\description{
Add a column called \code{method} which either will say SSD or AF.
}
\details{
The data is run through a set of rules to determine if a species
sensitivity distribution (SSD) or deterministic (DET) method can be used on
the data set.

Conditions for SSD:
\itemize{
\item number of fish species ≥ 3 AND
\item number of invertebrates ≥ 3 AND
\item number of plants or algae species ≥ 1
}

If these conditions are not met then a variation factor is the required.
}
\examples{
\dontrun{
data <- wqb_benchmark_method(data)
data <- wqb_benchmark_method(aggregated_data)
}
}
