% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-check-add-data.R
\name{wqb_check_add_data}
\alias{wqb_check_add_data}
\title{Check the Uploaded Data}
\usage{
wqb_check_add_data(data, template)
}
\arguments{
\item{data}{A data frame. The data you want to check.}

\item{template}{A data frame. The format the data should be in, in the
\href{https://poissonconsulting.github.io/chktemplate/}{chktemplate} format.}
}
\value{
A data frame
}
\description{
Checks the uploaded data for the basic requirements to ensure the data
matches the downloaded Ecotox data.
}
\details{
The values for the endpoint, trophic_group, and ecological_group
columns are checked against the data tables used to build the database. To
update the allowed values the corresponding csv file needs to be updated.
}
\examples{
\dontrun{
data <- wqb_check_add_data(data, template)
}
}
