% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-classify-duration.R
\name{wqb_classify_duration}
\alias{wqb_classify_duration}
\title{Classify Duration}
\usage{
wqb_classify_duration(data)
}
\arguments{
\item{data}{A data frame of the compiled data set.}
}
\value{
Invisible data frame
}
\description{
Each test will be classified as an acute or chronic.
}
\examples{
\dontrun{
classified_data <- wqb_classify_duration(compiled_data)
}
}
