% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-clean-data.R
\name{wqb_clean_data}
\alias{wqb_clean_data}
\title{Compile Data}
\usage{
wqb_clean_data(data, quiet = FALSE)
}
\arguments{
\item{data}{A data frame.}

\item{quiet}{Turn off message when quiet set to TRUE.}
}
\value{
Invisible data frame
}
\description{
Join database tables together and start filtering and cleaning data.
}
\details{
Check the resource document more details on the data added,
filter conditions and cleaning steps. This is part of Step 1.
}
\examples{
\dontrun{
data <- wqb_clean_data(data) 
}
}
