% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-compile-dataset.R
\name{wqb_compile_dataset}
\alias{wqb_compile_dataset}
\title{Compile the data set from the Database}
\usage{
wqb_compile_dataset(database)
}
\arguments{
\item{database}{A string to the location of the database.}
}
\value{
Invisible data frame
}
\description{
Compile data from the database to make the completed data set from the
external sources and US EPA ECOTOX database.
}
