% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-create-data-set.R
\name{wqb_create_data_set}
\alias{wqb_create_data_set}
\title{Create the Data Set for the App}
\usage{
wqb_create_data_set(
  file_path = "~/Ecotoxicology/ecotox",
  version = 1,
  folder_path = "~/Ecotoxicology/ecotox_db/",
  quiet = FALSE,
  ask = TRUE,
  save_rds = TRUE
)
}
\arguments{
\item{file_path}{A string of the file path location to save the downloaded
files. The default is your current working directory.}

\item{version}{An integer to indicate which version you want to download. The
default is 1 which downloads the most recent version.}

\item{folder_path}{Folder path to write to.}

\item{quiet}{Turn off message when quiet set to TRUE.}

\item{ask}{Turn off question when set to FALSE.}

\item{save_rds}{Saves the data set as a rds file in the folder_path.
Turn off when set to FALSE.}
}
\value{
A data frame
}
\description{
This function downloads the data, creates and adds all the data to the
database, compiles, standardizes, and classifies the data. The output
of this function is fed into the app.
}
\examples{
\dontrun{
 wqb_create_data_set(file_path = "~/Ecotoxicology/ecotox", 
 version = 1, 
 folder_path = "~/Ecotoxicology/ecotox_db/")
}
}
