% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-create-epa-ecotox.R
\name{wqb_create_epa_ecotox}
\alias{wqb_create_epa_ecotox}
\title{Create a SQLite/PostgreSQL Database from the US EPA ECOTOX data}
\usage{
wqb_create_epa_ecotox(file_path = ".")
}
\description{
Create a SQLite/PostgreSQL Database from the US EPA ECOTOX data
}
