% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-download-epa-ecotox.R
\name{wqb_download_epa_ecotox}
\alias{wqb_download_epa_ecotox}
\title{Download EPA ECOTOX Database Files}
\usage{
wqb_download_epa_ecotox(file_path = ".", version = 1)
}
\arguments{
\item{file_path}{A string of the file path location to save the downloaded
files. The default is your current working directory.}

\item{version}{An integer to indicate which version you want to download.
The default is 1 which downloads the most recent version.}
}
\value{
Invisible string of the file path the downloaded files were saved.
}
\description{
Download the EPA ECOTOX database files from their FTP site. The default is to
download the most recent version of the database.
}
\details{
You have the option of downloading older version of the database
but only up to the four most recent version. The most recent version is set
as 1 and the older version you can download is version 4.

The downloaded folder will contain various files that are needed to compile
the database.
}
\examples{
\dontrun{
wqb_download_epa_ecotox()

wqb_download_epa_ecotox("data_download")

# pull previous version of the database
wqb_download_epa_ecotox("data_download", version = 2)
}
}
