% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-filter-aquatic-tests.R
\name{wqb_filter_aquatic_tests}
\alias{wqb_filter_aquatic_tests}
\title{Filter Tests to Aquatic Species}
\usage{
wqb_filter_aquatic_tests(database)
}
\arguments{
\item{database}{A string to the location of the database.}
}
\value{
Invisible data frame
}
\description{
Filter to only tests on aquatic species.
}
\details{
The tests table is filtered down to only aquatic tests which is when the
\code{organism_habitat} column in the tests table is equal to "Water".

The output table is added to the database with the name
\code{tests_aquatic}.
}
\examples{
\dontrun{
wqb_filter_aquatic_tests(
  database = "ecotox_ascii_09_15_2022.sqlite"
)

wqb_filter_aquatic_tests(
  database = "ecotox_db/ecotox_ascii_09_15_2022.sqlite"
)
}
}
