% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-filter-chemical.R
\name{wqb_filter_chemical}
\alias{wqb_filter_chemical}
\title{Filter data set to a Single Chemical}
\usage{
wqb_filter_chemical(data, cas_num)
}
\arguments{
\item{data}{A data frame}

\item{cas_num}{A string of the cas number.}
}
\value{
Data frame
}
\description{
Filter data set to a Single Chemical
}
\examples{
\dontrun{
data <- wqb_filter_chemical(data, "129909906")
data <- wqb_filter_chemical(data, "1000984359")
}
}
