% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-generate-bench.R
\name{wqb_generate_bench}
\alias{wqb_generate_bench}
\title{Generate benchmark}
\usage{
wqb_generate_bench(data)
}
\arguments{
\item{data}{A data frame of clean and processed data filtered to only a
single chemical.}
}
\description{
Pass the data to generate the benchmark for DET or SSD method.
}
\examples{
\dontrun{
benchmark <- wqb_generate_bench(data)
}
}
