% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-generate-ctv.R
\name{wqb_generate_ctv}
\alias{wqb_generate_ctv}
\title{Generate Critical Toxicity Value}
\usage{
wqb_generate_ctv(data, dists = ssdtools::ssd_dists_bcanz())
}
\arguments{
\item{data}{A data frame}

\item{dists}{A character vector of the distributions to fit.}
}
\value{
A data frame
}
\description{
Determine the critical toxicity value for the species. The function will
apply the method listed in the method column.
}
\details{
The ctv_est_mg.L is the estimate of the critical toxicity value for
the species. The ctv_lcl_mg.L is the lower confidence limit. The ctv_ucl_mg.L
is the upper confidence limit.

The Deterministic method will always produce missing values in the
ctv_lcl_mg.L and ctv_ucl_mg.L columns.
}
\examples{
\dontrun{
ctv <- wqb_generate_ctv(data)
}
}
