% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-generate-det.R
\name{wqb_generate_det}
\alias{wqb_generate_det}
\title{Generate the Variation Factor Benchmark Guideline Value}
\usage{
wqb_generate_det(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame which contains the lowest concentration, the total
assessment factor and the benchmark value.
}
\description{
Finds the lowest concentration and then divides by the assessment
factor to output the benchmark value for the chemical.
}
\examples{
\dontrun{
bench_iodine <- wqb_generate_det(data)
}
}
