% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-generate-ssd.R
\name{wqb_generate_ssd}
\alias{wqb_generate_ssd}
\title{Generate the Benchmark Guideline Value using a SSD}
\usage{
wqb_generate_ssd(data)
}
\arguments{
\item{data}{}
}
\value{
A data frame which contains the HC5 value, the total
assessment factor and the benchmark value.
}
\description{
Generate the Benchmark Guideline Value using a SSD
}
\examples{
\dontrun{
 bench_iodine <- wqb_generate_ssd(data)
}
}
