% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-generate-ssd.R
\name{wqb_generate_ssd_fit}
\alias{wqb_generate_ssd_fit}
\title{Fit a SSD}
\usage{
wqb_generate_ssd_fit(data)
}
\arguments{
\item{data}{A data frame of clean and processed data filtered to only a
single chemical.}
}
\value{
A data frame
}
\description{
Fit a SSD
}
\examples{
\dontrun{
 fit <- wqb_generate_ssd_fit(data)
}
}
