% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-generate-ssd.R
\name{wqb_generate_ssd_hc5}
\alias{wqb_generate_ssd_hc5}
\title{Get a SSD HC5}
\usage{
wqb_generate_ssd_hc5(fit)
}
\arguments{
\item{data}{}
}
\value{
A data frame
}
\description{
Get a SSD HC5
}
\examples{
\dontrun{
 fit <- wqb_generate_ssd_fit(data)
}
}
