% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-generate-det.R
\name{wqb_method_det}
\alias{wqb_method_det}
\title{Deterministic Method to Generate Critical Toxicity Value for the Chemical}
\usage{
wqb_method_det(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame
}
\description{
Use the deterministic method to generate the critical toxicity value for the
data set.
}
\details{
Check the resource document for more details . This is Step 1.
}
\examples{
\dontrun{
bench_iodine <- wqb_method_det(data)
}
}
