% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-plot.R
\name{wqb_plot}
\alias{wqb_plot}
\title{Plots a single chemical}
\usage{
wqb_plot(data)
}
\arguments{
\item{data}{A data frame of clean and processed data filtered to only a
single chemical.}
}
\description{
Plots a single chemical
}
\examples{
\dontrun{
wqb_plot(data)
}
}
