% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-plot-ssd.R
\name{wqb_plot_ssd}
\alias{wqb_plot_ssd}
\title{Plot Species Sensitivity Distribution}
\usage{
wqb_plot_ssd(data, fit)
}
\arguments{
\item{data}{A data frame}

\item{fit}{The fit from ssd}
}
\description{
Plot results to see the species sensitivity distribution. The dashed line
shows the HC5 value.
}
\details{
This is a wrapper on \code{ssdtools::predict()} and
\code{ssdtools::ssd_plot()}.
}
\examples{
\dontrun{
wqb_plot_ssd(data, fit)
}
}
