% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-select-concentration-endpoints.R
\name{wqb_select_concentration_endpoints}
\alias{wqb_select_concentration_endpoints}
\title{Filter Results to Concentration Endpoint}
\usage{
wqb_select_concentration_endpoints(database)
}
\arguments{
\item{database}{A string to the location of the database.}
}
\value{
Invisible data frame
}
\description{
Determine the result ID's from concentration based endpoint tests. Logs the
concentration endpoints that are in the log format.
}
\details{
The concentration endpoints that have been selected are contained in
the extdata folder of the package in a file named
\code{concentration-endpoints.csv}. This file can be edited by adding or
deleting rows. Do not add new columns, rename columns or rename the file.
The file must contain the columns named \code{code} and \code{description}.

The full list of all possible endpoints is contained in the
\code{endpoint_codes} table in the ECOTOX downloaded data.

The output table that is written to the database is the results table
filtered to only contain the selected endpoints as listed in the
\code{concentration-endpoints.csv}.

The output table is added to the database with the name
\code{results_endpoint_concentration}.
}
\examples{
\dontrun{
wqb_select_concentration_endpoints(
  database = "ecotox_ascii_09_15_2022.sqlite"
)

wqb_select_concentration_endpoints(
  database = "ecotox_db/ecotox_ascii_09_15_2022.sqlite"
)
}
}
