% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-method-ssd.R
\name{wqb_ssd_fit}
\alias{wqb_ssd_fit}
\title{Fit BCANZ Distributions}
\usage{
wqb_ssd_fit(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame
}
\description{
Wrapper to \code{\link[ssdtools:ssd_fit_bcanz]{ssdtools::ssd_fit_bcanz()}}. The sp_aggre_conc_mg.L values are
the concentrations used.
}
\examples{
\dontrun{
fit <- wqb_ssd_fit(data)
}
}
