% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-method-ssd.R
\name{wqb_ssd_hc5}
\alias{wqb_ssd_hc5}
\title{Run SSD to get Hazard Concentrations}
\usage{
wqb_ssd_hc5(fit, nboot = 1000)
}
\arguments{
\item{fit}{The fit from ssd}

\item{nboot}{A count of the number of bootstrap samples to use to estimate
the SE and confidence limits. Default value of 1000.}
}
\value{
A data frame
}
\description{
Wrapper to the \code{\link[ssdtools:ssd_hc_bcanz]{ssdtools::ssd_hc_bcanz()}} function and selects only the row
which is 5\%.
}
\details{
The number of bootstrap samples is set to 1000 so the estimates are
quick to generate. Check out \code{\link[ssdtools:ssd_hc_bcanz]{ssdtools::ssd_hc_bcanz()}} for more details.
}
\examples{
\dontrun{
hc5 <- wqb_ssd_hc5(fit)
}
}
