% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-standardize-concentration.R
\name{wqb_standardize_concentration}
\alias{wqb_standardize_concentration}
\title{Add Concentration Unit Conversation Values for Standardizing the Data}
\usage{
wqb_standardize_concentration(database)
}
\arguments{
\item{database}{A string to the location of the database.}
}
\value{
Invisible data frame
}
\description{
Read in the concentration-conversion file and flag unit values that can be
converted and which ones will be removed. The column \code{conc_conversion_flag}
is added to the concentration_unit_codes table to indicate which units will
be kept and which will have the rows deleted. The
\code{conc_conversion_value_multiplier} column is added for each unit will be
converted to the unit specified in the \code{conc_conversion_unit} column.
}
\details{
The list of units to be converted are contained in a csv file in the
extdata folder of the package. The csv file can be edited by adding or
removing rows.  To add new rows get the \code{code} and \code{description} values
from the \code{concentration_unit_codes} table in the ECOTOX data and paste them
into the csv file.

Do not add new columns, rename columns or rename the file. The file must
only contain the columns: \code{code}, \code{description}, \code{conc_conversion_flag},
\code{conc_conversion_value_multiplier} and \code{conc_conversion_unit}.

The \code{code} values in the concentration-conversion file are matched to the
\code{code} values in the \code{concentration_unit_codes} table in the ECOTOX
downloaded data.

The \code{conc_conversion_value_multiplier} column contains the value need to
convert the unit into the unit listed in the \code{conc_conversion_unit} column.
}
\examples{
\dontrun{
concentration_unit_code_standardization <- wqb_standardize_concentration(
 database = "ecotox_ascii_09_15_2022.sqlite"
)

concentration_unit_code_standardization <- wqb_standardize_concentration(
 database = "ecotox_db/ecotox_ascii_09_15_2022.sqlite"
)
}
}
