% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-standardize-duration.R
\name{wqb_standardize_duration}
\alias{wqb_standardize_duration}
\title{Add Duration Unit Conversation Values for Standardizing the Data}
\usage{
wqb_standardize_duration(database)
}
\arguments{
\item{database}{A string to the location of the database.}
}
\value{
Invisible data frame
}
\description{
Read in the duration-conversion file and flag unit values that can be
converted and which ones will be removed. The column \code{duration_units_to_keep}
is added to the duration_unit_codes table to indicate which units will be
kept and which will have the rows deleted. The
\code{duration_value_multiplier_to_hours} column is added for each unit will be
converted to hours.
}
\details{
The list of units to be converted are contained in a csv file in the
extdata folder of the package. The csv file can be edited by adding or
removing rows.  To add new rows get the \code{code} and \code{description} values
from the \code{duration_unit_codes} table in the ECOTOX data and paste them into
the csv file.

Do not add new columns, rename columns or rename the file. The file must
only contain the columns: \code{code}, \code{description}, \code{duration_units_to_keep}
and \code{duration_value_multiplier_to_hours}.

The \code{code} values in the duration-conversion file are matched to the \code{code}
values in the \code{duration_unit_codes} table in the ECOTOX downloaded data.

The \code{duration_value_multiplier_to_hours} column contains the value need to
convert the unit into hours.
}
\examples{
\dontrun{
duration_unit_code_standardization <- wqb_standardize_duration(
 database = "ecotox_ascii_09_15_2022.sqlite"
)

duration_unit_code_standardization <- wqb_standardize_duration(
 database = "ecotox_db/ecotox_ascii_09_15_2022.sqlite"
)
}
}
