% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-standardize-effect.R
\name{wqb_standardize_effect}
\alias{wqb_standardize_effect}
\title{Standardize type of Effect}
\usage{
wqb_standardize_effect(data)
}
\arguments{
\item{data}{A data frame that is the output of the \code{wqb_classify_duration()}
function.}
}
\value{
Data frame
}
\description{
Standardize type of Effect
}
