% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-summary-af.R
\name{wqb_summary_af}
\alias{wqb_summary_af}
\title{Summary Table of the Assessment Factor Values}
\usage{
wqb_summary_af(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame
}
\description{
Summary Table of the Assessment Factor Values
}
\examples{
\dontrun{
summary_af <- wqb_summary_af(data)
}
}
