% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-summary-trophic.R
\name{wqb_summary_trophic_groups}
\alias{wqb_summary_trophic_groups}
\title{Summary Table of Assessment Factors Information}
\usage{
wqb_summary_trophic_groups(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame
}
\description{
Summary Table of Assessment Factors Information
}
\examples{
\dontrun{
summary_af <- wqb_summary_trophic_groups(data)
}
}
