% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-summary-trophic-species.R
\name{wqb_summary_trophic_species}
\alias{wqb_summary_trophic_species}
\title{Summary Table of Number of Species per Trophic Group}
\usage{
wqb_summary_trophic_species(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame.
}
\description{
Summary Table of Number of Species per Trophic Group
}
\examples{
\dontrun{
summary_trophic <- wqb_summary_trophic_species(data)
}
}
