% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqbench-package.R
\docType{package}
\name{wqbench-package}
\alias{wqbench}
\alias{wqbench-package}
\title{wqbench: Downloads and Compiles Data from the US EPA ECOTOX Database}
\description{
Download the most recent version of the US EPA Ecotox database to be ready to use. The data is then processed and compiled so it can be used to generate water quality guideline benchmarks for various chemicals and compounds.
}
\author{
\strong{Maintainer}: Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID})

}
\keyword{internal}
