% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-bc-species.R
\name{wqb_add_bc_species}
\alias{wqb_add_bc_species}
\title{Add BC Species to Database}
\usage{
wqb_add_bc_species(file_path, database)
}
\arguments{
\item{file_path}{A string of the file location for the file that contains
the list of BC Species.}

\item{database}{A string to the location of the database.}
}
\value{
Invisible data frame
}
\description{
Read in the list of BC species and then categorize each species as either
present in BC or not.
}
\details{
The BC species list must contain a single column called latin_name.
The latin_name column must consist of the genus and species separated by a
space.

The output table contains two columns: species_number and bc_species

The output table is added to the database with the name \code{species_bc_species}.
}
