% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-bc-species.R
\name{wqb_add_bc_species}
\alias{wqb_add_bc_species}
\title{Add BC Species to Database}
\usage{
wqb_add_bc_species(database)
}
\arguments{
\item{database}{A string to the location of the database.}
}
\value{
Invisible data frame
}
\description{
Read in the British Columbia species values and add them to the SQLite
database.
}
\details{
The BC species data is contained in a csv file in the extdata folder
of the package. This file can be edited by adding new species or removing
species. Do not add new columns, rename columns or rename the file. The
file must only contain a single column named \code{latin_name}.

The \code{latin_name} column must consist of the genus and species separated by
a space. The \code{latin_name} column in the bc-species file are matched to the
\code{latin_name} column in the species table of the ECOTOX downloaded data.

The output table that is written to the database contains two columns:
species_number and bc_species

The output table is added to the database with the name
\code{species_british_columbia}.
}
\examples{
\dontrun{
bc_species <- wqb_add_bc_species(
 database = "ecotox_ascii_09_15_2022.sqlite"
) 

bc_species <- wqb_add_bc_species(
 database = "ecotox_db/ecotox_ascii_09_15_2022.sqlite"
) 
}
}
