% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-bc-wqg.R
\name{wqb_add_bc_wqg}
\alias{wqb_add_bc_wqg}
\title{Add BC Water Quality Guidelines to Database}
\usage{
wqb_add_bc_wqg(database)
}
\arguments{
\item{database}{A string to the location of the database.}
}
\value{
Invisible data frame
}
\description{
Read in the British Columbia water quality guidelines (wqg) and add to the
chemicals tables in the database.
}
\details{
The BC wag data is stored in the BC Data Catalogue.

The \code{CAS_number} column in the bc wqg data are matched to the \code{cas_number}
column in the chemicals table of the ECOTOX downloaded data. A new column
\code{present_in_bc_wqg} is added to the chemicals table that codes each
chemical as either TRUE or FALSE.
}
\examples{
\dontrun{
chem_bc_wqg <- wqb_add_bc_wqg(
 database = "ecotox_ascii_09_15_2022.sqlite"
)

chem_bc_wqg <- wqb_add_bc_wqg(
 database = "ecotox_db/ecotox_ascii_09_15_2022.sqlite"
)
}
}
