% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-concentration-endpoints.R
\name{wqb_add_concentration_endpoints}
\alias{wqb_add_concentration_endpoints}
\title{Add the Concentration Endpoints to Database}
\usage{
wqb_add_concentration_endpoints(database)
}
\arguments{
\item{database}{A string to the location of the database.}
}
\value{
Invisible data frame
}
\description{
Read in the selected concentration endpoints and mark those values in the
endpoint_code table in the database by adding a \code{concentration_flag} column
to the table.
}
\details{
The list of concentration endpoints to be used is contained in a csv
file in the extdata folder of the package. The csv file can be edited by
adding or removing rows.  To add new rows get the \code{code} and \code{description}
values from the \code{endpoint_code} table in the ECOTOX data and paste them
into the csv file.

Do not add new columns, rename columns or rename the file. The file must
only contain the \code{code} and \code{description} column.

The \code{code} values in the endpoint-concentration file are matched to the
\code{code} values in the endpoint_code table in the ECOTOX downloaded data. Any
codes that are listed in the file are then coded as TRUE under a new column
called \code{concentration_flag} in the endpoint_table.
}
\examples{
\dontrun{
wqb_add_concentration_endpoints(
  database = "ecotox_ascii_09_15_2022.sqlite"
)

wqb_add_concentration_endpoints(
  database = "ecotox_db/ecotox_ascii_09_15_2022.sqlite"
)
}
}
