% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-concentration-endpoints.R
\name{wqb_add_concentration_endpoints}
\alias{wqb_add_concentration_endpoints}
\title{Add the Concentration Endpoints}
\usage{
wqb_add_concentration_endpoints(database, quiet = FALSE)
}
\arguments{
\item{database}{A string to the location of the database.}

\item{quiet}{Turn off message when quiet set to TRUE.}
}
\value{
Invisible data frame
}
\description{
Read in the concentration endpoints that have been selected to keep in the
final data set. Add a column to the endpoint_codes table in the database to
mark the corresponding endpoints.
}
\details{
The list of concentration endpoints is contained in a csv file in
the extdata folder of the package. The csv file can be edited by adding or
removing rows.  To add new rows get the \code{code} and \code{description} values
from the \code{endpoint_codes} table in the database and paste them into the csv
file.

Do not add new columns, rename columns or rename the file. The file must
only contain the \code{code} and \code{description} column.

The \code{code} values in the endpoint-concentration file are matched to the
\code{code} values in the endpoint_code table in the database. A new column
\code{concentration_flag} is added to the endpoint_code table that codes each
endpoint as TRUE if the endpoint is present in endpoint-concentration.csv
file.
}
\examples{
\dontrun{
wqb_add_concentration_endpoints(
  database = "ecotox_ascii_09_15_2022.sqlite"
)

wqb_add_concentration_endpoints(
  database = "ecotox_db/ecotox_ascii_09_15_2022.sqlite"
)
}
}
