% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-habitat.R
\name{wqb_add_habitat}
\alias{wqb_add_habitat}
\title{Code Media Type to Salt or Fresh Water by Species and Test}
\usage{
wqb_add_habitat(database)
}
\arguments{
\item{database}{A string to the location of the database.}
}
\value{
Invisible data frame
}
\description{
The media type code for each species and test has been coded to either salt water,
fresh water or not reported.
}
\details{
This function reads in the test table and codes each value to one
of three options: salt water, fresh water or not reported.

The output table is added to the database with the name
\code{species_tests_media}.
}
\examples{
\dontrun{
habitat_info <- wqb_add_habitat(
 database = "ecotox_ascii_09_15_2022.sqlite"
) 

habitat_info <- wqb_add_habitat(
 database = "ecotox_db/ecotox_ascii_09_15_2022.sqlite"
) 
}
}
