% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-lifestage.R
\name{wqb_add_lifestage}
\alias{wqb_add_lifestage}
\title{Add Life Stage Groups}
\usage{
wqb_add_lifestage(database, quiet = FALSE)
}
\arguments{
\item{database}{A string to the location of the database.}

\item{quiet}{Turn off message when quiet set to TRUE.}
}
\value{
Invisible data frame
}
\description{
Read in the life stage simple groups and add a column in the lifestage_code
table in the database to mark the corresponding values.
}
\details{
Only life stages related to fish or amphibians have been coded. The
purpose of the coding is to be able to simplify the many life stages into
three categories: els (early life stage), juveniles and adults. Not all
life stages have been coded into these three groups.

The life stage data is contained in a csv file in the extdata folder of the
package. The csv file can be edited by adding or removing rows.  To add new
rows get the \code{code} and \code{description} values from the \code{lifestage_code}
table in the dataset and paste them into the csv file and then add the
value to the \code{simple_lifestage} column.

Do not add new columns, rename columns or rename the file. The file must
only contain the \code{code}, \code{description_lifestage} and \code{simple_lifestage}
column.

The \code{code} values in the lifestage-codes.csv file are matched to the \code{code}
values in the lifestage_code table in the database. Any codes
that match are coded in a new column called \code{simple_lifestage}.
}
\examples{
\dontrun{
lifestage_codes <- wqb_add_lifestage(
  database = "ecotox_ascii_09_15_2022.sqlite"
)

lifestage_codes <- wqb_add_lifestage(
  database = "ecotox_db/ecotox_ascii_09_15_2022.sqlite"
)
}
}
