% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-add-trophic-group.R
\name{wqb_add_trophic_group}
\alias{wqb_add_trophic_group}
\title{Add Trophic Groups to Database}
\usage{
wqb_add_trophic_group(database, quiet = FALSE)
}
\arguments{
\item{database}{A string to the location of the database.}

\item{quiet}{Turn off message when quiet set to TRUE.}
}
\value{
Invisible data frame
}
\description{
Read in the trophic and ecological groups and and add a column to the
\code{species} table in the database that lists the groups for each species.
}
\details{
The trophic group data is contained in a csv file in the extdata
folder of the package. This file can be edited by adding or removing groups
and classes. Do not add new columns, rename columns or rename the file.

The trophic groups file must contain the columns: \code{class}, \code{order},
\code{ecological_group}, and \code{ecological_group_class}. The \code{class} and \code{order}
columns are matched to the \code{class} and \code{tax_order} columns in the \code{species}
table of the database and then adds the \code{ecological_group} and
\code{ecological_group_class} columns to the species table.
}
\examples{
\dontrun{
trophic_group <- wqb_add_trophic_group(
 database = "ecotox_ascii_09_15_2022.sqlite"
) 

trophic_group <- wqb_add_trophic_group(
 database = "ecotox_db/ecotox_ascii_09_15_2022.sqlite"
) 
}
}
