% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-af-bc-species.R
\name{wqb_af_bc_species}
\alias{wqb_af_bc_species}
\title{Determine the BC Species Assessment Factor}
\usage{
wqb_af_bc_species(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame
}
\description{
Determines an assessment factor based on the number of BC species present.
}
\details{
Check the resource document for how the assessment factor was
calculated.
}
\examples{
\dontrun{
data <- wqb_af_bc_species(data)
}
}
