% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-af-ecological.R
\name{wqb_af_ecological}
\alias{wqb_af_ecological}
\title{Determine the Ecological Assessment Factors}
\usage{
wqb_af_ecological(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame
}
\description{
Determines the assessment factor based on the presence of certain ecological
groups.
}
\details{
Check the resource document for how the assessment factor was
calculated.
}
\examples{
\dontrun{
data <- wqb_af_ecological(data)
}
}
