% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-af-variation.R
\name{wqb_af_variation}
\alias{wqb_af_variation}
\title{Determine the Species Variation Assessment Factor}
\usage{
wqb_af_variation(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame
}
\description{
Determine the assessment factor based on the number of trophic groups and
species.
}
\details{
Check the resource document for how the assessment factor was
calculated.
}
\examples{
\dontrun{
data <- wqb_af_variation(data)
}
}
