% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-af-variation.R
\name{wqb_af_variation}
\alias{wqb_af_variation}
\title{Determine the Assessment Factor}
\usage{
wqb_af_variation(data)
}
\arguments{
\item{data}{A data frame that is the output of the \code{wqb_benchmark-method()} function.}
}
\value{
A data frame
}
\description{
Based on a set of rules determine all the assessment factors for the
data set.
}
