% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-aggregate.R
\name{wqb_aggregate}
\alias{wqb_aggregate}
\title{Aggregate Data for each Species}
\usage{
wqb_aggregate(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame
}
\description{
Aggregate data to select the most sensitive value for each species. The
priority of effect level depends on the benchmark method that will be used.
}
\details{
Check the resource document for the rules used to aggregate the data
down to a single value per species. This is Step 5.
}
\examples{
\dontrun{
data <- wqb_aggregate(data)
}
}
