% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-benchmark-method.R
\name{wqb_benchmark_method}
\alias{wqb_benchmark_method}
\title{Determine Benchmark Method}
\usage{
wqb_benchmark_method(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame
}
\description{
Determine if a species sensitivity distribution (SSD) or deterministic method
will be used to determine the aquatic life water quality benchmark value
for the data. A column called \code{method} will be added to the data indicating
which method will be used.
}
\details{
Check the resource document for the rules used to determine which
benchmark method will be used. This is Step 4.
}
\examples{
\dontrun{
data <- wqb_benchmark_method(data)
data <- wqb_benchmark_method(aggregated_data)
}
}
