% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-classify-duration.R
\name{wqb_classify_duration}
\alias{wqb_classify_duration}
\title{Classify Duration}
\usage{
wqb_classify_duration(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame
}
\description{
Classify each test as either acute or chronic in the \code{duration_class}
column.
}
\details{
Check the resource document for the rules used to determine the
classification for each trophic group. This is Step 2.
}
\examples{
\dontrun{
classified_data <- wqb_classify_duration(compiled_data)
}
}
