% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-classify-duration.R
\name{wqb_classify_duration}
\alias{wqb_classify_duration}
\title{Classify Duration}
\usage{
wqb_classify_duration(data)
}
\arguments{
\item{data}{A data frame of the compiled data set.}
}
\value{
Invisible data frame
}
\description{
Each test will be classified as an acute or chronic and a new column called
\code{duration_class} will be added to the data set.
}
\details{
Classifications for acute and chronic are separated by trophic group
(\code{ecological_group}). The values are classified as acute, chronic in a new
column called \code{duration_class}.

Values not meeting this criteria are classified as not specified and are
removed.

Fish and amphibians
Acute:
- duration ≤ 96 hours
Chronic:
- life stage (\code{simple_lifestage}) is juvenile or adult and
duration ≥ 504 hours
- life stage (\code{simple_lifestage}) is early life stages (els) and
duration ≥ 168 hours

Aquatic invertebrates
Acute:
- duration ≤ 96 hours
Chronic:
- Planktonic invertebrates (\code{ecological_group_class}) and duration > 96 hours
- Regular invertebrates (\code{ecological_group_class}) and duration ≥ 168 hours

Algae
Acute:
- duration ≤ 24 hours
Chronic:
- duration > 24 hours

Aquatic Plants
Acute:
- duration ≤ 48 hours
Chronic:
\itemize{
\item duration > 7 days (used Lemna protocol as a guide, EC 2007)
}

Reference
Environment Canada 2007. Test for measuring the inhibition of growth using
the freshwater macrophyte, Lemna minor.  Available online at:
https://publications.gc.ca/collections/collection_2013/ec/En49-7-1-37-eng.pdf
}
\examples{
\dontrun{
classified_data <- wqb_classify_duration(compiled_data)
}
}
