% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-compile-data.R
\name{wqb_compile_dataset}
\alias{wqb_compile_dataset}
\title{Compile Data}
\usage{
wqb_compile_dataset(database, quiet = FALSE)
}
\arguments{
\item{database}{A string to the location of the database.}

\item{quiet}{Turn off message when quiet set to TRUE.}
}
\value{
Invisible data frame
}
\description{
Join database tables together and start filtering and cleaning data.
}
\details{
Check the resource document more details on the data added,
filter conditions and cleaning steps. This is Step 1.
}
\examples{
\dontrun{
data_compiled <- wqb_compile_dataset(
 database = "ecotox_ascii_09_15_2022.sqlite"
) 

data_compiled <- wqb_compile_dataset(
 database = "ecotox_db/ecotox_ascii_09_15_2022.sqlite"
) 
}
}
