% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-create-epa-ecotox.R
\name{wqb_create_epa_ecotox}
\alias{wqb_create_epa_ecotox}
\title{Create the ECOTOX Database}
\usage{
wqb_create_epa_ecotox(folder_path = ".", data_path)
}
\arguments{
\item{folder_path}{Folder path to write to.}

\item{data_path}{Folder path to the downloaded ECOTOX folder}
}
\value{
Invisible string of the file path of the database.
}
\description{
Create a SQLite database from the US EPA ECOTOX downloaded files.
}
\details{
This functions reads in the text files in the folder and writes
them to the database.
}
\examples{
\dontrun{
wqb_create_epa_ecotox(data_path = "ecotox_ascii_12_15_2022")
}
}
