% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-generate-ssd.R
\name{wqb_generate_ssd_hc5}
\alias{wqb_generate_ssd_hc5}
\title{Get a SSD HC5}
\usage{
wqb_generate_ssd_hc5(fit, nboot = 100)
}
\arguments{
\item{fit}{The fit from ssd}

\item{nboot}{A count of the number of bootstrap samples to use to estimate
the se and confidence limits. A value of 10000 is recommended for official
guidelines.}
}
\value{
A data frame
}
\description{
Get a SSD HC5
}
\examples{
\dontrun{
 fit <- wqb_generate_ssd_hc5(data)
}
}
