% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-method-ssd.R
\name{wqb_method_ssd}
\alias{wqb_method_ssd}
\title{SSD Method to Calculate Critical Toxicity Value for the Chemical}
\usage{
wqb_method_ssd(data, fit, nboot = 1000)
}
\arguments{
\item{data}{A data frame}

\item{fit}{The fit}

\item{nboot}{The number of bootstrap samples. Default value of 1000.}
}
\value{
A data frame
}
\description{
Use a species sensitivity distribution to calculate the critical toxicity
value for the data set. The critical toxicity value is the hazardous
concentration for 5\% of species (HC5) when the SSD method is used.
}
\details{
A wrapper on the ssdtools package function
\code{\link[ssdtools:ssd_hc_bcanz]{ssdtools::ssd_hc_bcanz()}} that only returns the HC5 concentration.
}
\examples{
\dontrun{
hc5 <- wqb_method_ssd(data, fit)
}
}
