% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-method-ssd.R
\name{wqb_method_ssd}
\alias{wqb_method_ssd}
\title{Generate the Benchmark Guideline Value using a SSD}
\usage{
wqb_method_ssd(data, fit)
}
\arguments{
\item{data}{A data frame of clean and processed data filtered to only a
single chemical.}

\item{fit}{The fit from ssd}
}
\value{
A data frame which contains the Critical Toxicity Value (HC5) and
upper and lower bound.
}
\description{
Generate the Benchmark Guideline Value using a SSD
}
\examples{
\dontrun{
 bench_iodine <- wqb_method_ssd(data, fit)
}
}
