% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-plot.R
\name{wqb_plot}
\alias{wqb_plot}
\title{Plot Data}
\usage{
wqb_plot(data)
}
\arguments{
\item{data}{A data frame}
}
\description{
Plot the standardized data to see the concentrations per species for each
endpoint.
}
\examples{
\dontrun{
wqb_plot(data)
}
}
