% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-plot.R
\name{wqb_plot}
\alias{wqb_plot}
\title{Plots a single chemical}
\usage{
wqb_plot(data, y_axis = "conc1_mean_std_effect")
}
\arguments{
\item{data}{A data frame of clean and processed data filtered to only a
single chemical.}

\item{y_axis}{A string of the column to use for the y-axis}
}
\value{
A ggplot object
}
\description{
Plots a single chemical
}
\examples{
\dontrun{
wqb_plot(data)
}
}
