% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-plot-det.R
\name{wqb_plot_det}
\alias{wqb_plot_det}
\title{Plot the Deterministic Method Results}
\usage{
wqb_plot_det(data)
}
\arguments{
\item{data}{A data frame}
}
\description{
Plot deterministic results to see the standardized data, critical toxicity
value and aquatic life benchmark.
}
\examples{
\dontrun{
wqb_plot_det(data)
}
}
