% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-plot-ssd.R
\name{wqb_plot_ssd}
\alias{wqb_plot_ssd}
\title{Plot SSD}
\usage{
wqb_plot_ssd(data, fit)
}
\arguments{
\item{data}{A data frame of clean and processed data filtered to only a
single chemical.}

\item{fit}{The fit from ssd}
}
\description{
Plot SSD
}
