% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-select-concentration-endpoints.R
\name{wqb_select_concentration_endpoints}
\alias{wqb_select_concentration_endpoints}
\title{Select only Concentration Related Results}
\usage{
wqb_select_concentration_endpoints(file_path, database)
}
\description{
Determine the result ID's from concentration based endpoint tests.
}
