% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-method-ssd.R
\name{wqb_ssd_hc5}
\alias{wqb_ssd_hc5}
\title{Run SSD to get Hazard Concentrations}
\usage{
wqb_ssd_hc5(fit, nboot = 10000)
}
\arguments{
\item{fit}{The fit from ssd}

\item{nboot}{A count of the number of bootstrap samples to use to estimate
the se and confidence limits. A value of 10000 is recommended for official
guidelines.}
}
\value{
A data frame
}
\description{
Wrapper to the \code{ssdtools::ssd_hc_bcanz()} function and selects only the
row which is 5\%.
}
\examples{
\dontrun{
hc5 <- wqb_ssd_hc5(fit)
}
}
