% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqb-standardize-effect.R
\name{wqb_standardize_effect}
\alias{wqb_standardize_effect}
\title{Standardize type of Effect}
\usage{
wqb_standardize_effect(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame
}
\description{
Determine and apply factor needed to standardize the endpoints.
}
\details{
Check the resource document for details the rules used to determine
standardized the endpoints. This is Step 3.
}
\examples{
\dontrun{
standardized_effect_data <- wqb_standardize_effect(data)
standardized_effect_data <- wqb_standardize_effect(classified_data)
}
}
