% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqbench-package.R
\docType{package}
\name{wqbench-package}
\alias{wqbench}
\alias{wqbench-package}
\title{wqbench: Calculate Aquatic Life Benchmarks}
\description{
Download recent versions of the US EPA ECOTOX database. Clean, standardize and classify data so values are comparable. Use a SSD or deterministic method to determine the critical toxicity value and assessment factors to compute the aquatic life water quality benchmark for a compound.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bcgov.github.io/wqbench/}
}

}
\author{
\strong{Maintainer}: Angeline Tillmanns \email{Angeline.Tillmanns@gov.bc.ca}

Authors:
\itemize{
  \item Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID})
}

Other contributors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID}) [contributor]
  \item Andy Teucher \email{andy@andyteucher.ca} (\href{https://orcid.org/0000-0002-7840-692X}{ORCID}) [contributor]
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
